Fix Linux CI: rewrite apply_ilu on inspector-executor + u32/i32 casts
Linux bindgen output differs from Windows in two ways that broke CI:

- `mkl_dcsrtrsv` (legacy CSR triangular solve) is not exposed in the
  Linux bindings, so `apply_ilu` now drives two `mkl_sparse_d_trsv`
  calls through a temporary `mkl_sparse_d_create_csr` handle. Same
  semantics (unit-lower L solve, then non-unit-upper U solve), just
  via the modern Inspector-Executor API that ships on both platforms.

- Several MKL enum constants (sparse_status_t, mkl_jit_status_t, the
  VSL_CONV_MODE_* family) come back typed as u32 on Linux but i32 on
  Windows. Match arms and error-path casts now bridge the two without
  relying on either typing.

+++ onemkl/src/blas/jit.rs
@@ -49,10 +49,15 @@ pub enum JitStatus {
 impl JitStatus {
     #[inline]
     fn from_raw(status: sys::mkl_jit_status_t::Type) -> Result<Self> {
+        // Compare against the bindgen-typed constants directly so
+        // this works on both Linux (where the type is u32) and
+        // Windows (where it's i32).
         match status {
-            s if s == sys::mkl_jit_status_t::MKL_JIT_SUCCESS as c_int => Ok(Self::Compiled),
-            s if s == sys::mkl_jit_status_t::MKL_NO_JIT as c_int => Ok(Self::Fallback),
-            other => Err(Error::LapackComputationFailure { info: other }),
+            s if s == sys::mkl_jit_status_t::MKL_JIT_SUCCESS => Ok(Self::Compiled),
+            s if s == sys::mkl_jit_status_t::MKL_NO_JIT => Ok(Self::Fallback),
+            other => Err(Error::LapackComputationFailure {
+                info: other as i32,
+            }),
         }
     }
 }
